+++ src/main/java/controller/PaymentController.java
@@ -2,9 +2,9 @@ package controller;
 
 import domain.Table;
 import domain.TableRepository;
-import domain.payment.ChickenDiscountPolicy;
 import domain.payment.Payment;
 import domain.payment.PaymentType;
+import domain.payment.discount.ChickenDiscountPolicy;
 import view.InputView;
 import view.OutputView;
 
@@ -18,6 +18,7 @@ public class PaymentController implements PosController {
 		PaymentType paymentType = PaymentType.of(InputView.inputPaymentId());
 		Payment payment = new Payment(new ChickenDiscountPolicy(), paymentType, table);
 		OutputView.printTotalMoney(payment.calculate());
+
 		table.clear();
 	}
 }
