+++ src/main/java/domain/payment/discount/BeverageDiscountPolicy.java
@@ -1,0 +1,5 @@
+package domain.payment.discount;
+
+public class BeverageDiscountPolicy {
+	// TODO: 2020/04/10 추가 구현
+}
